git-svn-id: https://source.openmpt.org/svn/openmpt/branches/OpenMPT-1.28@11038 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/PatternFindReplace.cpp
@@ -429,7 +429,7 @@ void CViewPattern::OnEditFindNext()
 						else if(FindReplace::instance.replaceParamAction == FindReplace::ReplaceValue)
 							param = paramReplace;
 
-						m->SetValueVolCol(static_cast<uint16>(Clamp(param, 0, ModCommand::maxColumnValue)));
+						m->SetValueVolCol(static_cast<uint16>(Clamp(param, 0, static_cast<int>(ModCommand::maxColumnValue))));
 					}
 
 					if(FindReplace::instance.replaceFlags[FindReplace::PCValue])
@@ -443,7 +443,7 @@ void CViewPattern::OnEditFindNext()
 						else if(FindReplace::instance.replaceVolumeAction == FindReplace::ReplaceValue)
 							value = valueReplace;
 
-						m->SetValueEffectCol(static_cast<uint16>(Clamp(value, 0, ModCommand::maxColumnValue)));
+						m->SetValueEffectCol(static_cast<uint16>(Clamp(value, 0, static_cast<int>(ModCommand::maxColumnValue))));
 					}
 
 					SetModified(false);
